+++ src/test/scala/com/guizmaii/distances/implementations/google/geocoder/GoogleGeocoderSpec.scala
@@ -66,7 +66,7 @@ class GoogleGeocoderSpec extends WordSpec with Matchers with ScalaFutures with B
       def testGeocoder(postalCode: PostalCode, place: LatLong): Assertion = {
         cache.flushAll()
         cache.get(postalCode) shouldBe None
-        whenReady(geocoder.geocodeT(postalCode).runAsync) { result =>
+        whenReady(geocoder.geocodePostalCodeT(postalCode).runAsync) { result =>
           result shouldBe place
           cache.get(postalCode) shouldBe Some(place)
         }
@@ -91,7 +91,7 @@ class GoogleGeocoderSpec extends WordSpec with Matchers with ScalaFutures with B
         cache.flushAll()
         cache.set(postalCode)(place)
         cache.get(postalCode) shouldBe Some(place)
-        geocoder.geocodeT(postalCode).runAsync.futureValue shouldBe place
+        geocoder.geocodePostalCodeT(postalCode).runAsync.futureValue shouldBe place
       }
       "just return" should {
         "Lille" in {